+++ apis/cassandra/v1beta1/cassandradatacenter_types.go
@@ -52,6 +52,10 @@ const (
 	// CassandraDatacenter is deleted.
 	DecommissionOnDeleteAnnotation = "cassandra.datastax.com/decommission-on-delete"
 
+	// NoAutomatedCleanUpAnnotation prevents the cass-operator from creating a CassandraTask to do cleanup after the
+	// cluster has gone through scale up operation.
+	NoAutomatedCleanupAnnotation = "cassandra.datastax.com/no-cleanup"
+
 	CassNodeState = "cassandra.datastax.com/node-state"
 
 	ProgressUpdating ProgressState = "Updating"
@@ -407,8 +411,8 @@ type CassandraDatacenterStatus struct {
 
 // CassandraDatacenter is the Schema for the cassandradatacenters API
 // +k8s:openapi-gen=true
-//+kubebuilder:object:root=true
-//+kubebuilder:subresource:status
+// +kubebuilder:object:root=true
+// +kubebuilder:subresource:status
 // +kubebuilder:resource:path=cassandradatacenters,scope=Namespaced,shortName=cassdc;cassdcs
 type CassandraDatacenter struct {
 	metav1.TypeMeta   `json:",inline"`