+++ tests/scale_up/scale_up_suite_test.go
@@ -70,7 +70,7 @@ var _ = Describe(testName, func() {
 			ns.WaitForDatacenterReady(dcName)
 
 			// Ensure we have a single CassandraTask created which is a cleanup (and it succeeded)
-			ns.CheckForCompletedCassandraTasks(dcName, "cleanup", 1)
+			ns.WaitForCompletedCassandraTasks(dcName, "cleanup", 1)
 
 			step = "scale up to 4 nodes"
 			json = "{\"spec\": {\"size\": 4}}"
@@ -83,7 +83,7 @@ var _ = Describe(testName, func() {
 			ns.WaitForDatacenterCondition(dcName, "ScalingUp", string(corev1.ConditionFalse))
 
 			// Ensure we have two CassandraTasks created which are cleanup (and they succeeded)
-			ns.CheckForCompletedCassandraTasks(dcName, "cleanup", 2)
+			ns.WaitForCompletedCassandraTasks(dcName, "cleanup", 2)
 
 			step = "scale up to 5 nodes"
 			json = "{\"spec\": {\"size\": 5}}"
@@ -96,7 +96,7 @@ var _ = Describe(testName, func() {
 			ns.WaitForDatacenterCondition(dcName, "ScalingUp", string(corev1.ConditionFalse))
 
 			// Ensure we have three CassandraTasks created which are cleanup (and they succeeded)
-			ns.CheckForCompletedCassandraTasks(dcName, "cleanup", 3)
+			ns.WaitForCompletedCassandraTasks(dcName, "cleanup", 3)
 
 			step = "check recorded host IDs"
 			nodeStatusesHostIds := ns.GetNodeStatusesHostIds(dcName)